prebas {Rprebas}	R Documentation

PREBAS forest model

Description

This function runs PREBAS model.

Usage

prebas(nYears, pCROBAS = pCROB, pPRELES = pPREL, etmodel = 0, pYASSO = pYAS, pAWEN = parsAWEN, siteInfo = NA, initVar = NA,
  thinning=NA,initClearcut = c(1.5,0.5,0.0431969,0.,0.),PAR,TAir,VPD,Precip,CO2, P0=NA,  soilC = NA,
  weatherYasso = NA, litterSize = NA, soilCtot = numeric(nYears), defaultThin = 1.,
  ClCut = 1., inDclct = NA, inAclct = NA, yassoRun = 0)

Arguments

nYears
Number of years to run the model

pCROBAS
Matrix of parameter sets, each column corresponds to a species. Default values pCROBAS = pCROB are the parameter sets for Scots pine, Norway spruce and Silver birch. print(pCROB) to see the parameter values and names.

pPRELES
Parameter vector for PRELES model. Default values (pPRELES = pPREL) is the calibration for European Boreal forests (Minunno et al. 2016)

etmodel
Evapotranspiration model for PRELES. Default etmodel = 0. Possible values -1, 0, 1, 2

pYASSO
Parameter vector for YASSO model. Default pYASSO = pYAS (Liski et al., 2017)

pAWEN
Matrix of parameter sets for partitioning tree organs (foliage, branch and woody) in AWEN pools. Columns referes to different species. Note that the columns of pAWEN must correspond to the species of pPRELES. Default pAWEN = parsAWEN are for Scots pine, Norway spruce and Silver birch.

siteInfo
Vector of site info SiteID, climID, siteType. Default = c(1,1,3), i.e. siteType = 3.

initVar
Matrix with initial stand values for all the tree strata. Columns correspond to the layers in the stand. Initial information needed are: SpeciesID (a number corresponding to the species parameter values of pPRELES columns), Age (years), average height of the layer (H, m), average diameter at breast height of the layer (D, cm), basal area of the layer (BA, m2 ha-1), average height of the crown base of the layer (Hc, m). If initVar is not provided the model is initialized from plantation using default planting parameters (see initClearcut) and assuming that Pine, Spruce and Birch are equally present at plantation.

thinning
A matrix with thinnig inputs. Rows correspond to a thinning event. Column 1 year from the start of the simulation; column 2 is siteID; column 3 layer where thinnings are carried out; column 4 to 7 stand variables (H, D, B, Hc); column 8 parameter that indicates if the stand variables (column 4:7) are provided as fraction of the actual model outputs. (see examples)

initClearcut
A numeric vector with initial stand variables after clearcut: H, D, BA, Hc, Ainit. Ainit is the year when the stand reaches the default values = c(1.5,0.5,0.0431969,0.,0.) and is automatically computed using air temperature.

PAR
A numeric vector of daily sums of photosynthetically active radiation, mmol/m2.

TAir
A numeric vector of daily mean temperature, degrees C.

VPD
A numeric vector of daily mean vapour pressure deficits, kPa.

Precip
A numeric vector of daily rainfall, mm

CO2
A numeric vector of air CO2, ppm

P0
A numeric vector with the annual potential photosynthesis (gC m-2 y-1). If P0 is not provided PRELES is used to compute P0 using fAPAR = 1.

soilC
Initial soil carbon compartments for each layer. Array with dimentions = c(nYears,5,3,nLayers). The second dimention (5) corresponds to the AWENH pools; the third dimention (3) corresponds to the tree organs (foliage, branch and stem).

weatherYasso
Annual weather inputs for Yasso.

litterSize
Marix with litter inputs for YASSO. Rows are tree organs, columns correspond to the layers.

soilCtot
Vector with total initial soil carbon

defaultThin
If defaultThin = 1 (default) Finnish standard managment practices are applied (ref).

ClCut
If ClCut = 1 clearcuts are applied. If inDclct = NA and inAclct = NA Finnish standard clearcut practices are applied (ref).

inDclct
Vector of Diameter (cm) threshold for clearcut. Each element correspond to a layer of the stand, if only one value is provided the same value is applied to all the layers. The different elements of the vector are for the different layers. The dominant species (highest basal area) is considered for clearcut.

inAclct
Vector of Age (year) threshold for clearcut. Each element correspond to a layer of the stand, if only one value is provided the same value is applied to all the layers. The different elements of the vector are for the different layers. The dominant species (highest basal area) is considered for clearcut.

yassoRun
If yassoRun=1 the YASSO model is run to compute the carbon balance of the soil.

Value

output
An array with annual model outputs. 1st dimension corresponds to the number of years of the simulation (nYears); 2nd dimension corresponds to the output variables (see list below); 3rd dimension corresponds to the number of layers in the stand (nLayers); 4th dimensions reports the state of the stand (1) and (2) the variables of the harvested trees (2).

Output variables: 1."siteID"
2."climID"
3."sitetype"
4."species"
5."ETS" effective temperature sums
6."P0" Potential annual gross primary production (gC m-2 y-1)
7."age" Age of the layer (years)
8."Respi_m /10000." (kgC m-2 y-1)
9."Respi_tot" Autotrophic respiration (gC m-2 y-1)
10."GPP/1000" Total GPP (kgC ha-1 y-1)
11."H" Layer average height (m)
12."D" Layer average diameter at breast height (cm)
13."BA" Layer basal area (m-2 ha-1)
14."Hc_base" Base of crown height (m)
15."Cw" Crorn width (m)
16."Lc" Length of the crown (m)
17."N" Layer density
18."npp" net primary production (gC m-2 y-1)
19."leff" Effective leaf area
20."keff" Effective light extintion coefficient
21."lproj" Projected leaf area
22."ET_preles" Annual evapotranspiration (mm m-2)
23."weight" Layer weight on photosynthesis
24."Wbranch" Branch biomass (kgC ha-1)
25."WfineRoots" Fine roots biomass (kgC ha-1)
26."Litter_fol" Foliage litter (kgC ha-1)
27."Litter_fr" Fine root litter (kgC ha-1)
28."Litter_branch" Branch litter (kgC ha-1)
29."Litter_wood" Woody litter (kgC ha-1)
30."V" Layer volume (m3 ha-1)
31."Wstem" Stem Biomass (kgC ha-1)
32."W_croot" Course root Biomass (kgC ha-1)
33."wf_STKG" Foliage biomass (kgC ha-1)
34."wf_treeKG" Foliage biomass of the average tree (kgC ha-1)
35."B_tree" Basal area of average tree (m2 ha-1)
36."Light" light interseption
37."Vharvested" harvested volume (m3 ha-1)
38."Vtot" total volume of the Layer considering also dead trees and harvested volume (m3 ha-1)
39."soilC" totaal soil carbon (kgC ha-1)
40."aSW" average available soil water (mm m-2)
41."summerSW" summer soil water (mm m-2)
42."Vmort" volume of dead trees (m3 ha-1)
43."gross growth" (m3 ha-1 y-1)
44."GPPspecies" Gross primary production per layer (gC m-2 y-1)
45."Rh species" (gC m-2 y-1)
46."NEP sp" Net ecosystem exchange (gC m-2 y-1)

dailyPRELES
Matrix with daily output from preles. 1st column is the daily GPP (gC m-2 d-1), 2nd column daily evapotranspiration (mm m-2), 3rd column daily soil water (mm -2).

Author(s)

Francesco Minunno

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (nYears, pCROBAS = pCROB, pPRELES = pPREL, pYASSO = pYAS,
    siteInfo = NA, thinning = matrix(0, 2, 10), nThinning = max(2,
        nrow(thinning)), initClearcut = c(1.5, 0.5, 0.0431969,
        0, 0), PAR, TAir, VPD, Precip, CO2, P0 = NA, initVar = NA,
    soilC = array(0, dim = c(nYears, 5, 3, 3)), weatherYasso = matrix(0,
        nYears, 3), litterSize = matrix(0, 3, 3), soilCtot = numeric(nYears),
    defaultThin = 1, ClCut = 1, inDclct = NA, inAclct = NA, yassoRun = 0)
{
    nLayers <- ifelse(!is.na(initVar), ncol(initVar), 3)
    nSp = ncol(pCROBAS)
    if (is.na(siteInfo))
        siteInfo = c(1, 1, 3)
    varNam <- getVarNam()
    nVar <- length(varNam)
    nLayers <- nSp <- ncol(pCROBAS)
    layerNam <- paste("layer", 1:nSp)
    output <- array(0, dim = c((nYears), nVar, nSp, 2), dimnames = list(NULL,
        varNam, layerNam, c("stand", "thinned")))
    fAPAR <- rep(0.7, nYears)
    Temp <- TAir[1:(365 * nYears)] - 5
    ETS <- pmax(0, Temp, na.rm = T)
    ETS <- matrix(ETS, 365, nYears)
    ETS <- colSums(ETS)
    if (is.na(P0)) {
        P0 <- PRELES(DOY = rep(1:365, nYears), PAR = PAR, TAir = TAir,
            VPD = VPD, Precip = Precip, CO2 = CO2, fAPAR = rep(1,
                length(PAR)), LOGFLAG = 0, p = pPRELES)$GPP
        P0 <- matrix(P0, 365, nYears)
        P0 <- colSums(P0)
    }
    ETSthres <- 1000
    ETSmean <- mean(ETS)
    if (any(!is.na(c(inDclct, inAclct)))) {
        if (is.na(inDclct))
            inDclct <- 9999999.99
        if (is.na(inAclct))
            inAclct <- 9999999.99
    }
    if (ClCut == 1 & is.na(initVar) & is.na(inDclct))
        inDclct <- c(ClCutD_Pine(ETSmean, ETSthres, siteInfo[3]),
            ClCutD_Spruce(ETSmean, ETSthres, siteInfo[3]), ClCutD_Birch(ETSmean,
                ETSthres, siteInfo[3]))
    if (ClCut == 1 & is.na(initVar) & is.na(inAclct))
        inAclct <- c(ClCutA_Pine(ETSmean, ETSthres, siteInfo[3]),
            ClCutA_Spruce(ETSmean, ETSthres, siteInfo[3]), ClCutA_Birch(ETSmean,
                ETSthres, siteInfo[3]))
    if (any(is.na(inDclct)))
        inDclct[is.na(inDclct)] <- 9999999.99
    if (length(inDclct) == 1)
        inDclct <- rep(inDclct, nSp)
    if (any(is.na(inAclct)))
        inAclct[is.na(inAclct)] <- 9999999.99
    if (length(inAclct) == 1)
        inAclct <- rep(inAclct, nSp)
    if (is.na(initVar)) {
        initVar <- matrix(NA, 6, nSp)
        initVar[1, ] <- 1:nSp
        initVar[3, ] <- initClearcut[1]
        initVar[4, ] <- initClearcut[2]
        initVar[5, ] <- initClearcut[3]/nSp
        initVar[6, ] <- initClearcut[4]
    }
    xx <- min(10, nYears)
    Ainit = 6 + 2 * 3.5 - 0.005 * (sum(ETS[1:xx])/xx) + 2.25
    initVar[2, which(is.na(initVar[2, ]))] <- initClearcut[5] <- round(Ainit)
    weatherPreles <- array(c(PAR, TAir, VPD, Precip, CO2), dim = c(365,
        nYears, 5))
    weatherPreles <- aperm(weatherPreles, c(2, 1, 3))
    prebas <- .Fortran("prebas", nYears = as.integer(nYears),
        nLayers = as.integer(nLayers), nSp = as.integer(nSp),
        siteInfo = as.numeric(siteInfo), pCROBAS = as.matrix(pCROBAS),
        initVar = as.matrix(initVar), thinning = as.matrix(thinning),
        output = as.array(output), nThinning = as.integer(nThinning),
        maxYearSite = as.integer(nYears), fAPAR = as.numeric(fAPAR),
        initClearcut = as.numeric(initClearcut), ETS = as.numeric(ETS),
        P0 = as.numeric(P0), weather = as.array(weatherPreles),
        DOY = as.integer(1:365), pPRELES = as.numeric(pPRELES),
        soilC = as.array(soilC), pYASSO = as.numeric(pYASSO),
        weatherYasso = as.matrix(weatherYasso), litterSize = as.matrix(litterSize),
        soilCtot = as.numeric(soilCtot), defaultThin = as.double(defaultThin),
        ClCut = as.double(ClCut), inDclct = as.double(inDclct),
        inAclct = as.double(inAclct), yassoRun = as.double(yassoRun))
    class(prebas) <- "prebas"
    return(prebas)
  }

[Package Rprebas version 0.1.0 Index]